requirements.txt
torch==2.2.2
torchvision==0.17.2

fastapi==0.110.0
uvicorn==0.27.1

pillow==10.2.0
numpy==1.26.4


mlflow
dvc

opencv-python
scikit-learn
matplotlib
seaborn
pandas
python-multipart
pytest==8.1.1
